--- src/main/java/racing_car_with_winner/exception/InvalidInputException.java
@@ -1,7 +1,0 @@
-package racing_car_with_winner.exception;
-
-public class InvalidInputException extends RuntimeException {
-    public InvalidInputException(String message) {
-        super(message);
-    }
-}
