--- src/main/java/racing_car_with_winner/model/CarFactory.java
@@ -1,22 +1,0 @@
-package racing_car_with_winner.model;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import racing_car_with_winner.view.CarNameInput;
-
-public class CarFactory {
-    private final CarNameInput carNameInput;
-
-    public CarFactory(final CarNameInput carNameInput) {
-        this.carNameInput = carNameInput;
-    }
-
-    public List<Car> build() {
-        return Arrays.stream(carNameInput.parse())
-                     .map(CarName::new)
-                     .map(Car::new)
-                     .collect(Collectors.toList());
-    }
-}
