--- src/main/java/racing_car_with_winner/model/MovableStrategy.java
@@ -1,5 +1,0 @@
-package racing_car_with_winner.model;
-
-public interface MovableStrategy {
-    boolean canMove();
-}
