--- src/main/java/racing_car_with_winner/view/CarNameInput.java
@@ -1,14 +1,0 @@
-package racing_car_with_winner.view;
-
-public class CarNameInput {
-    private static final String DELIMITER = ",";
-    private final String input;
-
-    public CarNameInput(final String input) {
-        this.input = input;
-    }
-
-    public String[] parse() {
-        return input.split(DELIMITER);
-    }
-}
