--- src/main/java/racing_car_with_winner/view/UserInputInterface.java
@@ -1,6 +1,0 @@
-package racing_car_with_winner.view;
-
-public interface UserInputInterface {
-    void show(String message);
-    String getStringInput();
-}
