--- src/main/java/racing_car_with_winner/view/UserOutputInterface.java
@@ -1,5 +1,0 @@
-package racing_car_with_winner.view;
-
-public interface UserOutputInterface {
-    public void show(String message);
-}
