+++ src/main/java/racingcar/winner/exception/InvalidInputException.java
@@ -1,0 +1,7 @@
+package racingcar.winner.exception;
+
+public class InvalidInputException extends RuntimeException {
+    public InvalidInputException(String message) {
+        super(message);
+    }
+}
