+++ src/main/java/racingcar/winner/model/GameRule.java
@@ -1,0 +1,9 @@
+package racingcar.winner.model;
+
+public final class GameRule {
+    private static final int MOVABLE_CRITERIA = 4;
+
+    public static boolean canMove(int value) {
+        return value >= MOVABLE_CRITERIA;
+    }
+}
