+++ src/main/java/racingcar/winner/model/MovableStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.winner.model;
+
+public interface MovableStrategy {
+    boolean canMove();
+}
