+++ src/main/java/racingcar/winner/view/ConsoleUserInputInterface.java
@@ -1,0 +1,17 @@
+package racingcar.winner.view;
+
+import java.util.Scanner;
+
+public class ConsoleUserInputInterface implements UserInputInterface {
+    Scanner scanner = new Scanner(System.in);
+
+    @Override
+    public void show(String message) {
+        System.out.print(message);
+    }
+
+    @Override
+    public String getStringInput() {
+        return scanner.nextLine();
+    }
+}
