+++ src/main/java/racingcar/winner/view/ConsoleUserOutputInterface.java
@@ -1,0 +1,8 @@
+package racingcar.winner.view;
+
+public class ConsoleUserOutputInterface implements UserOutputInterface {
+    @Override
+    public void show(String message) {
+        System.out.print(message);
+    }
+}
