+++ src/main/java/racingcar/winner/view/UserInputInterface.java
@@ -1,0 +1,6 @@
+package racingcar.winner.view;
+
+public interface UserInputInterface {
+    void show(String message);
+    String getStringInput();
+}
