+++ src/main/java/racingcar/winner/view/UserOutputInterface.java
@@ -1,0 +1,5 @@
+package racingcar.winner.view;
+
+public interface UserOutputInterface {
+    public void show(String message);
+}
